Symptom 2
Input: test-file2.md
Expected Output: [https://something.com, some-page.html]

Actual Output: Looping between -1 and 42. Symptom 1 looped between multiple indexs for open brackets.

Suspected Reason: I believe that this is due to a lack of a statement to catch what happens when markdown.indexOf does not return a non-negative index. This causes it too loop as it sets the index to -1, which allows for it to essentially go over values it has already gone over.