Symptom 2 Fix
For the moment, Ive temporarily removed the checker for images that I believed was causing Symptom 1. I created checks that break the while loop if any of the indexes of both paren and brackets returned a negative value. I know that if they returned a negative value, it signified they had reached the end of the list and were unable to go any further.