Symptom 3
Input: test-file5.md
Expected Output: []

Actual Output: [page.com]

Suspected Reason: I believe that this error is caused due to there not being a check if the data within parenthesis are valid links in markdown format. That is to say, the current code interprets anything in parenthesis, whether it is in a valid position adjacent to a bracket pair or in an invalid position not adjacent to a bracket pair.