Symptom 3 Fix
My hunch was right and it seems that a lack of a check for the brackets and parenthesis being adjacent to one another was the culprit. I've added an if statement that checks if the char to the right of the closeBracket is an open parenthesis. If it is, that means the following parenthesis pair is valid and the link is added. If not, then the parenthesis pair are not a valid link given it is not adjacent to a brackets pair and is not added.