Symptom 1 Fix
Symptom 1 was initially caused by a faulty image link if statement to check if the following link was for an image or for a url. What happened originally was that the statement would not update the index correctly, causing the index to decrease, reseting the loop and causing an infinite loop where index would never reach the end of the while condition. With the current code, two conditions must be met, which is that the index of openBracket is greater than 0 and that the char at the index openBracket -1 is equal to exclamation point. In this scenario, the method will skip that link and go to the next one, updating the currentIndex accordingly.